+++ packages/block-editor-utils/src/helpers/getControlFields.ts
@@ -33,11 +33,13 @@ function getControlFields(
     // Set default field by merging both blockAttributes meta and editorFields hints.
     if (fieldConfig) {
       fields.push({
+        ...fieldConfig,
         name: key,
         label: fieldConfig.label ?? key,
         type: fieldType,
         location: fieldConfig.location ?? 'editor',
         control: fieldConfig?.control ?? control,
+        options: fieldConfig?.options ?? [],
       });
     } else {
       // Set default field by using only blockAttributes meta
@@ -47,6 +49,7 @@ function getControlFields(
         type: fieldType,
         location: 'editor',
         control,
+        options: [],
       });
     }
   });